+++ sync-for-payables/Codat/Sync/Payables/Models/Components/BankAccountMappingOption.cs
@@ -79,7 +79,7 @@ namespace Codat.Sync.Payables.Models.Components
         /// </remarks>
         /// </summary>
         [JsonProperty("accountType")]
-        public AccountType? AccountType { get; set; }
+        public BankAccountType? AccountType { get; set; }
 
         [JsonProperty("sourceModifiedDate")]
         public string? SourceModifiedDate { get; set; }