+++ sync-for-payables/RELEASES.md
@@ -149,3 +149,13 @@ Based on:
 - [csharp v7.0.0] sync-for-payables
 ### Releases
 - [NuGet v7.0.0] https://www.nuget.org/packages/Codat.Sync.Payables/7.0.0 - sync-for-payables
+
+## 2024-10-30 13:00:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.423.3 (2.442.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v8.0.0] sync-for-payables
+### Releases
+- [NuGet v8.0.0] https://www.nuget.org/packages/Codat.Sync.Payables/8.0.0 - sync-for-payables